requirements.txt
boto3==1.4.7
botocore==1.7.39
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.6
django-cors-headers==2.1.0
django-filter==1.0.4
django-paypal==0.4.1
django-storages==1.6.5
django-templated-mail==1.0.0
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
djoser==1.1.1
docutils==0.14
futures==3.1.1
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
mercadopago==0.3.4
netifaces==0.10.7
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
PyJWT==1.5.3
python-dateutil==2.6.1
python-http-client==3.0.0
pytz==2017.2
requests==2.18.4
s3transfer==0.1.11
sendgrid==5.2.0
six==1.11.0
sorl-thumbnail==12.3
stripe==1.67.0
urllib3==1.22
whitenoise==3.2
